+++ test/localdatetime_test.dart
@@ -29,7 +29,7 @@ Future setup() async {
 @Test()
 void ToDateTimeUnspecified()
 {
-  LocalDateTime zoned = new LocalDateTime.at(2011, 3, 5, 1, 0);
+  LocalDateTime zoned = new LocalDateTime.at(2011, 3, 5, 1, 0, 0);
   DateTime expected = new DateTime(2011, 3, 5, 1, 0, 0); //, DateTimeKind.Unspecified);
   DateTime actual = zoned.toDateTimeLocal();
   expect(expected, actual);
@@ -46,7 +46,7 @@ void ToDateTimeUnspecified()
 void ToDateTimeUnspecified_TruncatesTowardsStartOfTime(int year)
 {
   if (Platform.isWeb) {
-    var ldt = new LocalDateTime.at(year, 1, 1, 13, 15, seconds: 55).plusNanoseconds(
+    var ldt = new LocalDateTime.at(year, 1, 1, 13, 15, 55).plusNanoseconds(
         TimeConstants.nanosecondsPerSecond - 1); //.PlusNanoseconds(TimeConstants.nanosecondsPerSecond - 1);
     var expected = new DateTime(year, 1, 1, 13, 15, 55 /*, DateTimeKind.Unspecified*/)
         .add(new Duration(milliseconds: TimeConstants.millisecondsPerSecond - 1));
@@ -54,7 +54,7 @@ void ToDateTimeUnspecified_TruncatesTowardsStartOfTime(int year)
     expect(actual, expected);
   }
   else {
-    var ldt = new LocalDateTime.at(year, 1, 1, 13, 15, seconds: 55).plusNanoseconds(
+    var ldt = new LocalDateTime.at(year, 1, 1, 13, 15, 55).plusNanoseconds(
         TimeConstants.nanosecondsPerSecond - 1); //.PlusNanoseconds(TimeConstants.nanosecondsPerSecond - 1);
     var expected = new DateTime(year, 1, 1, 13, 15, 55 /*, DateTimeKind.Unspecified*/)
         .add(new Duration(microseconds: TimeConstants.microsecondsPerSecond - 1));
@@ -75,7 +75,7 @@ void ToDateTimeUnspecified_OutOfRange()
 @Test()
 void FromDateTime()
 {
-  LocalDateTime expected = new LocalDateTime.at(2011, 08, 18, 20, 53);
+  LocalDateTime expected = new LocalDateTime.at(2011, 08, 18, 20, 53, 0);
   // for (DateTimeKind kind in Enum.GetValues(typeof(DateTimeKind)))
   DateTime x = new DateTime.utc(2011, 08, 18, 20, 53, 0); //, kind);
   LocalDateTime actual = new LocalDateTime.fromDateTime(x);
@@ -86,7 +86,7 @@ void FromDateTime()
 void FromDateTime_WithCalendar()
 {
   // Julian calendar is 13 days behind Gregorian calendar in the 21st century
-  LocalDateTime expected = new LocalDateTime.at(2011, 08, 05, 20, 53, calendar: CalendarSystem.julian);
+  LocalDateTime expected = new LocalDateTime.at(2011, 08, 05, 20, 53, 0, calendar: CalendarSystem.julian);
 // print('Expected day of year: ${expected.date.DaysSinceEpoch}');
 
   // todo: I don't understand what the test is doing here, this doesn't work with DateTime() local.
@@ -102,7 +102,7 @@ void FromDateTime_WithCalendar()
 void TimeProperties_AfterEpoch()
 {
   // Use the largest valid year as part of validating against overflow
-  LocalDateTime ldt = new LocalDateTime.at(GregorianYearMonthDayCalculator.maxGregorianYear, 1, 2, 15, 48, seconds: 25).plusNanoseconds(123456789);
+  LocalDateTime ldt = new LocalDateTime.at(GregorianYearMonthDayCalculator.maxGregorianYear, 1, 2, 15, 48, 25).plusNanoseconds(123456789);
   expect(15, ldt.hour);
   expect(3, ldt.clockHourOfHalfDay);
   expect(48, ldt.minute);
@@ -124,7 +124,7 @@ void TimeProperties_AfterEpoch()
 void TimeProperties_BeforeEpoch()
 {
   // Use the smallest valid year number as part of validating against overflow
-  LocalDateTime ldt = new LocalDateTime.at(GregorianYearMonthDayCalculator.minGregorianYear, 1, 2, 15, 48, seconds: 25).plusNanoseconds(123456789);
+  LocalDateTime ldt = new LocalDateTime.at(GregorianYearMonthDayCalculator.minGregorianYear, 1, 2, 15, 48, 25).plusNanoseconds(123456789);
   expect(15, ldt.hour);
   expect(3, ldt.clockHourOfHalfDay);
   expect(48, ldt.minute);
@@ -159,7 +159,7 @@ void DateTime_Roundtrip_OtherCalendarInBcl()
 @Test()
 void WithCalendar()
 {
-  LocalDateTime isoEpoch = new LocalDateTime.at(1970, 1, 1, 0, 0);
+  LocalDateTime isoEpoch = new LocalDateTime.at(1970, 1, 1, 0, 0, 0);
   LocalDateTime julianEpoch = isoEpoch.withCalendar(CalendarSystem.julian);
   expect(1969, julianEpoch.year);
   expect(12, julianEpoch.month);
@@ -171,7 +171,7 @@ void WithCalendar()
 @Test()
 void TimeOfDay_Before1970()
 {
-  LocalDateTime dateTime = new LocalDateTime.at(1965, 11, 8, 12, 5, seconds: 23);
+  LocalDateTime dateTime = new LocalDateTime.at(1965, 11, 8, 12, 5, 23);
   LocalTime expected = new LocalTime(12, 5, 23);
   expect(expected, dateTime.time);
 }
@@ -180,7 +180,7 @@ void TimeOfDay_Before1970()
 @Test()
 void TimeOfDay_After1970()
 {
-  LocalDateTime dateTime = new LocalDateTime.at(1975, 11, 8, 12, 5, seconds: 23);
+  LocalDateTime dateTime = new LocalDateTime.at(1975, 11, 8, 12, 5, 23);
   LocalTime expected = new LocalTime(12, 5, 23);
   expect(expected, dateTime.time);
 }
@@ -189,7 +189,7 @@ void TimeOfDay_After1970()
 @Test()
 void Date_Before1970()
 {
-  LocalDateTime dateTime = new LocalDateTime.at(1965, 11, 8, 12, 5, seconds: 23);
+  LocalDateTime dateTime = new LocalDateTime.at(1965, 11, 8, 12, 5, 23);
   LocalDate expected = new LocalDate(1965, 11, 8);
   expect(expected, dateTime.date);
 }
@@ -198,7 +198,7 @@ void Date_Before1970()
 @Test()
 void Date_After1970()
 {
-  LocalDateTime dateTime = new LocalDateTime.at(1975, 11, 8, 12, 5, seconds: 23);
+  LocalDateTime dateTime = new LocalDateTime.at(1975, 11, 8, 12, 5, 23);
   LocalDate expected = new LocalDate(1975, 11, 8);
   expect(expected, dateTime.date);
 }
@@ -207,7 +207,7 @@ void Date_After1970()
 void DayOfWeek_AroundEpoch()
 {
   // Test about couple of months around the Unix epoch. If that works, I'm confident the rest will.
-  LocalDateTime dateTime = new LocalDateTime.at(1969, 12, 1, 0, 0);
+  LocalDateTime dateTime = new LocalDateTime.at(1969, 12, 1, 0, 0, 0);
   for (int i = 0; i < 60; i++)
   {
     // Check once per hour of the day, just in case something's messed up based on the time of day.
@@ -222,27 +222,27 @@ void DayOfWeek_AroundEpoch()
 @Test()
 void ClockHourOfHalfDay()
 {
-  expect(12, new LocalDateTime.at(1975, 11, 8, 0, 0).clockHourOfHalfDay);
-  expect(1, new LocalDateTime.at(1975, 11, 8, 1, 0).clockHourOfHalfDay);
-  expect(12, new LocalDateTime.at(1975, 11, 8, 12, 0).clockHourOfHalfDay);
-  expect(1, new LocalDateTime.at(1975, 11, 8, 13, 0).clockHourOfHalfDay);
-  expect(11, new LocalDateTime.at(1975, 11, 8, 23, 0).clockHourOfHalfDay);
+  expect(12, new LocalDateTime.at(1975, 11, 8, 0, 0, 0).clockHourOfHalfDay);
+  expect(1, new LocalDateTime.at(1975, 11, 8, 1, 0, 0).clockHourOfHalfDay);
+  expect(12, new LocalDateTime.at(1975, 11, 8, 12, 0, 0).clockHourOfHalfDay);
+  expect(1, new LocalDateTime.at(1975, 11, 8, 13, 0, 0).clockHourOfHalfDay);
+  expect(11, new LocalDateTime.at(1975, 11, 8, 23, 0, 0).clockHourOfHalfDay);
 }
 
 @Test()
 void Operators_SameCalendar()
 {
-  LocalDateTime value1 = new LocalDateTime.at(2011, 1, 2, 10, 30);
-  LocalDateTime value2 = new LocalDateTime.at(2011, 1, 2, 10, 30);
-  LocalDateTime value3 = new LocalDateTime.at(2011, 1, 2, 10, 45);
+  LocalDateTime value1 = new LocalDateTime.at(2011, 1, 2, 10, 30, 0);
+  LocalDateTime value2 = new LocalDateTime.at(2011, 1, 2, 10, 30, 0);
+  LocalDateTime value3 = new LocalDateTime.at(2011, 1, 2, 10, 45, 0);
   TestHelper.TestOperatorComparisonEquality(value1, value2, [value3]);
 }
 
 @Test()
 void Operators_DifferentCalendars_Throws()
 {
-  LocalDateTime value1 = new LocalDateTime.at(2011, 1, 2, 10, 30);
-  LocalDateTime value2 = new LocalDateTime.at(2011, 1, 3, 10, 30, calendar: CalendarSystem.julian);
+  LocalDateTime value1 = new LocalDateTime.at(2011, 1, 2, 10, 30, 0);
+  LocalDateTime value2 = new LocalDateTime.at(2011, 1, 3, 10, 30, 0, calendar: CalendarSystem.julian);
 
   expect(value1 == value2, isFalse);
   expect(value1 != value2, isTrue);
@@ -256,9 +256,9 @@ void Operators_DifferentCalendars_Throws()
 @Test()
 void CompareTo_SameCalendar()
 {
-  LocalDateTime value1 = new LocalDateTime.at(2011, 1, 2, 10, 30);
-  LocalDateTime value2 = new LocalDateTime.at(2011, 1, 2, 10, 30);
-  LocalDateTime value3 = new LocalDateTime.at(2011, 1, 2, 10, 45);
+  LocalDateTime value1 = new LocalDateTime.at(2011, 1, 2, 10, 30, 0);
+  LocalDateTime value2 = new LocalDateTime.at(2011, 1, 2, 10, 30, 0);
+  LocalDateTime value3 = new LocalDateTime.at(2011, 1, 2, 10, 45, 0);
 
   expect(value1.compareTo(value2), 0);
   expect(value1.compareTo(value3),  lessThan(0));
@@ -272,8 +272,8 @@ void CompareTo_DifferentCalendars_Throws()
   dynamic IslamicEpoch = null;
 
   CalendarSystem islamic = CalendarSystem.getIslamicCalendar(IslamicLeapYearPattern.Base15, IslamicEpoch.Astronomical);
-  LocalDateTime value1 = new LocalDateTime.at(2011, 1, 2, 10, 30);
-  LocalDateTime value2 = new LocalDateTime.at(1500, 1, 1, 10, 30, calendar: islamic);
+  LocalDateTime value1 = new LocalDateTime.at(2011, 1, 2, 10, 30, 0);
+  LocalDateTime value2 = new LocalDateTime.at(1500, 1, 1, 10, 30, 0, calendar: islamic);
 
   expect(() => value1.compareTo(value2), throwsArgumentError);
 // expect(() => ((IComparable)value1).CompareTo(value2), throwsArgumentError);
@@ -283,9 +283,9 @@ void CompareTo_DifferentCalendars_Throws()
 @Test()
 void IComparableCompareTo_SameCalendar()
 {
-  LocalDateTime value1 = new LocalDateTime.at(2011, 1, 2, 10, 30);
-  LocalDateTime value2 = new LocalDateTime.at(2011, 1, 2, 10, 30);
-  LocalDateTime value3 = new LocalDateTime.at(2011, 1, 2, 10, 45);
+  LocalDateTime value1 = new LocalDateTime.at(2011, 1, 2, 10, 30, 0);
+  LocalDateTime value2 = new LocalDateTime.at(2011, 1, 2, 10, 30, 0);
+  LocalDateTime value3 = new LocalDateTime.at(2011, 1, 2, 10, 45, 0);
 
   Comparable i_value1 = /*(IComparable)*/value1;
   Comparable i_value3 = /*(IComparable)*/value3;
@@ -299,7 +299,7 @@ void IComparableCompareTo_SameCalendar()
 @Test()
 void IComparableCompareTo_Null_Positive()
 {
-  var instance = new LocalDateTime.at(2012, 3, 5, 10, 45);
+  var instance = new LocalDateTime.at(2012, 3, 5, 10, 45, 0);
   Comparable i_instance = /*(IComparable)*/instance;
   Object arg = null;
   var result = i_instance.compareTo(arg);
@@ -311,7 +311,7 @@ void IComparableCompareTo_Null_Positive()
 @Test()
 void IComparableCompareTo_WrongType_ArgumentException()
 {
-  var instance = new LocalDateTime.at(2012, 3, 5, 10, 45);
+  var instance = new LocalDateTime.at(2012, 3, 5, 10, 45, 0);
   Comparable i_instance = /*(IComparable)*/instance;
   var arg = new LocalDate(2012, 3, 6);
   expect(() => i_instance.compareTo(arg), willThrow<TypeError>());
@@ -321,7 +321,7 @@ void IComparableCompareTo_WrongType_ArgumentException()
 void WithOffset()
 {
   var offset = new Offset.fromHoursAndMinutes(5, 10);
-  var localDateTime = new LocalDateTime.at(2009, 12, 22, 21, 39, seconds: 30);
+  var localDateTime = new LocalDateTime.at(2009, 12, 22, 21, 39, 30);
   var offsetDateTime = localDateTime.withOffset(offset);
   expect(localDateTime, offsetDateTime.localDateTime);
   expect(offset, offsetDateTime.offset);
@@ -330,7 +330,7 @@ void WithOffset()
 @Test()
 void InUtc()
 {
-  var local = new LocalDateTime.at(2009, 12, 22, 21, 39, seconds: 30);
+  var local = new LocalDateTime.at(2009, 12, 22, 21, 39, 30);
   var zoned = local.inUtc();
   expect(local, zoned.localDateTime);
   expect(Offset.zero, zoned.offset);
@@ -341,7 +341,7 @@ void InUtc()
 @Test()
 void InZoneStrictly_InWinter()
 {
-  var local = new LocalDateTime.at(2009, 12, 22, 21, 39, seconds: 30);
+  var local = new LocalDateTime.at(2009, 12, 22, 21, 39, 30);
   var zoned = local.inZoneStrictly(Pacific);
   expect(local, zoned.localDateTime);
   expect(new Offset.fromHours(-8), zoned.offset);
@@ -350,7 +350,7 @@ void InZoneStrictly_InWinter()
 @Test()
 void InZoneStrictly_InSummer()
 {
-  var local = new LocalDateTime.at(2009, 6, 22, 21, 39, seconds: 30);
+  var local = new LocalDateTime.at(2009, 6, 22, 21, 39, 30);
   var zoned = local.inZoneStrictly(Pacific);
   expect(local, zoned.localDateTime);
   expect(new Offset.fromHours(-7), zoned.offset);
@@ -361,7 +361,7 @@ void InZoneStrictly_InSummer()
 @Test()
 void InZoneStrictly_ThrowsWhenAmbiguous()
 {
-  var local = new LocalDateTime.at(2009, 11, 1, 1, 30);
+  var local = new LocalDateTime.at(2009, 11, 1, 1, 30, 0);
   expect(() => local.inZoneStrictly(Pacific), willThrow<AmbiguousTimeError>());
 }
 
@@ -370,7 +370,7 @@ void InZoneStrictly_ThrowsWhenAmbiguous()
 @Test()
 void InZoneStrictly_ThrowsWhenSkipped()
 {
-  var local = new LocalDateTime.at(2009, 3, 8, 2, 30);
+  var local = new LocalDateTime.at(2009, 3, 8, 2, 30, 0);
   expect(() => local.inZoneStrictly(Pacific), willThrow<SkippedTimeError>());
 }
 
@@ -379,7 +379,7 @@ void InZoneStrictly_ThrowsWhenSkipped()
 @Test()
 void InZoneLeniently_AmbiguousTime_ReturnsEarlierMapping()
 {
-  var local = new LocalDateTime.at(2009, 11, 1, 1, 30);
+  var local = new LocalDateTime.at(2009, 11, 1, 1, 30, 0);
   var zoned = local.inZoneLeniently(Pacific);
   expect(local, zoned.localDateTime);
   expect(new Offset.fromHours(-7), zoned.offset);
@@ -391,9 +391,9 @@ void InZoneLeniently_AmbiguousTime_ReturnsEarlierMapping()
 @Test()
 void InZoneLeniently_ReturnsStartOfSecondInterval()
 {
-  var local = new LocalDateTime.at(2009, 3, 8, 2, 30);
+  var local = new LocalDateTime.at(2009, 3, 8, 2, 30, 0);
   var zoned = local.inZoneLeniently(Pacific);
-  expect(new LocalDateTime.at(2009, 3, 8, 3, 30), zoned.localDateTime);
+  expect(new LocalDateTime.at(2009, 3, 8, 3, 30, 0), zoned.localDateTime);
   expect(new Offset.fromHours(-7), zoned.offset);
 }
 
@@ -401,8 +401,8 @@ void InZoneLeniently_ReturnsStartOfSecondInterval()
 void InZone()
 {
   // Don't need much for this - it only delegates.
-  var ambiguous = new LocalDateTime.at(2009, 11, 1, 1, 30);
-  var skipped = new LocalDateTime.at(2009, 3, 8, 2, 30);
+  var ambiguous = new LocalDateTime.at(2009, 11, 1, 1, 30, 0);
+  var skipped = new LocalDateTime.at(2009, 3, 8, 2, 30, 0);
   expect(new ZonedDateTime.atLeniently(ambiguous, Pacific), ambiguous.inZone(Pacific, Resolvers.lenientResolver));
   expect(new ZonedDateTime.atLeniently(skipped, Pacific), skipped.inZone(Pacific, Resolvers.lenientResolver));
 }
@@ -413,7 +413,7 @@ void DefaultConstructor()
 {
   // todo: LocalDateTime()
   var actual = new LocalDateTime(new LocalDate(1, 1, 1), new LocalTime(0, 0, 0));
-  expect(new LocalDateTime.at(1, 1, 1, 0, 0), actual);
+  expect(new LocalDateTime.at(1, 1, 1, 0, 0, 0), actual);
 }
 
 //@Test()
@@ -448,8 +448,8 @@ void DefaultConstructor()
 @Test()
 void MinMax_DifferentCalendars_Throws()
 {
-  LocalDateTime ldt1 = new LocalDateTime.at(2011, 1, 2, 2, 20);
-  LocalDateTime ldt2 = new LocalDateTime.at(1500, 1, 1, 5, 10, calendar: CalendarSystem.julian);
+  LocalDateTime ldt1 = new LocalDateTime.at(2011, 1, 2, 2, 20, 0);
+  LocalDateTime ldt2 = new LocalDateTime.at(1500, 1, 1, 5, 10, 0, calendar: CalendarSystem.julian);
 
   expect(() => LocalDateTime.max(ldt1, ldt2), throwsArgumentError);
   expect(() => LocalDateTime.min(ldt1, ldt2), throwsArgumentError);
@@ -458,8 +458,8 @@ void MinMax_DifferentCalendars_Throws()
 @Test()
 void MinMax_SameCalendar()
 {
-  LocalDateTime ldt1 = new LocalDateTime.at(1500, 1, 1, 7, 20, calendar: CalendarSystem.julian);
-  LocalDateTime ldt2 = new LocalDateTime.at(1500, 1, 1, 5, 10, calendar: CalendarSystem.julian);
+  LocalDateTime ldt1 = new LocalDateTime.at(1500, 1, 1, 7, 20, 0, calendar: CalendarSystem.julian);
+  LocalDateTime ldt2 = new LocalDateTime.at(1500, 1, 1, 5, 10, 0, calendar: CalendarSystem.julian);
 
   expect(ldt1, LocalDateTime.max(ldt1, ldt2));
   expect(ldt1, LocalDateTime.max(ldt2, ldt1));