+++ test/timezones/algiers_test.dart
@@ -41,7 +41,7 @@ void GetPeriod_BeforeLast()
 @Test()
 void GetPeriod_AfterLastTransition()
 {
-  var may1981 = new ZonedDateTime.atStrictly(new LocalDateTime.at(1981, 5, 1, 0, 0, seconds: 1), DateTimeZone.utc).toInstant();
+  var may1981 = new ZonedDateTime.atStrictly(new LocalDateTime.at(1981, 5, 1, 0, 0, 1), DateTimeZone.utc).toInstant();
   var actual = Algiers.getZoneInterval(may1981);
   var expected = IZoneInterval.newZoneInterval("CET", new Instant.fromUnixTimeMicroseconds(357523200000000), null, new Offset(TimeConstants.secondsPerHour), Offset.zero);
   expect(expected, actual);