+++ test/timezones/zoneinterval_test.dart
@@ -42,8 +42,8 @@ void PassthroughProperties()
 @Test()
 void ComputedProperties()
 {
-  LocalDateTime start = new LocalDateTime.at(2011, 6, 3, 19, 15);
-  LocalDateTime end = new LocalDateTime.at(2011, 8, 2, 22, 45);
+  LocalDateTime start = new LocalDateTime.at(2011, 6, 3, 19, 15, 0);
+  LocalDateTime end = new LocalDateTime.at(2011, 8, 2, 22, 45, 0);
   expect(start, SampleInterval.isoLocalStart);
   expect(end, SampleInterval.isoLocalEnd);
   expect(SampleEnd - SampleStart, SampleInterval.totalTime);